Update mainTurbCorrect.m
In File mainTurbCorrect.m when loadInput() function is called there is an error in the number of input parameters. There should be four input parameters whereas when calling the loadInput function five parameters are set as input. 

So I removed the startFrame from inputparameters list. 
I changed the line 53 of mainTurbCorrect.m from 
[input, inputU, inputV] = loadInput(dirname, extfile,startFrame, totalFrame, resizeRatio);

to

[input, inputU, inputV] = loadInput(dirname, extfile,totalFrame, resizeRatio);

Now it works fine.